+++ examples/pokemon_explorer/lib/src/client.dart
@@ -1,9 +1,8 @@
 import 'package:gql_http_link/gql_http_link.dart';
 import 'package:ferry/ferry.dart';
 import 'package:ferry_hive_store/ferry_hive_store.dart';
-import 'package:hive/hive.dart';
-import 'package:hive_flutter/hive_flutter.dart';
-import 'package:pokemon_explorer/__generated__/schema.schema.gql.dart';
+import 'package:hive_flutter/adapters.dart';
+import 'package:pokemon_explorer/main.dart';
 
 Future<Client> initClient() async {
   await Hive.initFlutter();
@@ -16,7 +15,7 @@ Future<Client> initClient() async {
 
   final cache = Cache(store: store);
 
-  final link = HttpLink("https://pokeapi.dev");
+  final link = HttpLink(apiUrl);
 
   final client = Client(
     link: link,